requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
